+++ fancy_core/GpuRingBuffer.h
@@ -14,16 +14,16 @@ namespace Fancy
 
     void Create(const GpuBufferProperties& someParameters, GpuResoruceLockOption aLockOption, const void* pInitialData = nullptr);
 
-    uint GetFreeDataSize() const;
-    bool AllocateAndWrite(const void* someData, uint aDataSize, uint& anOffsetOut);
-    bool Allocate(uint aDataSize, uint& anOffsetOut);
+    uint64 GetFreeDataSize() const;
+    bool AllocateAndWrite(const void* someData, uint64 aDataSize, uint64& anOffsetOut);
+    bool Allocate(uint64 aDataSize, uint64& anOffsetOut);
     void Reset() { myOffset = 0u; }
     const GpuBuffer* GetBuffer() const { return myBuffer.get(); }
 
   protected:
     SharedPtr<GpuBuffer> myBuffer;
     uint8* myData;
-    uint myOffset;
+    uint64 myOffset;
     GpuResoruceLockOption myLockType;
   };
 //---------------------------------------------------------------------------//
